// src/auth/role.enum.ts
export enum Role {
    AdminNegocio = 'adminNegocio',
    LecturaNegocio = 'lecturaNegocio',
    EscrituraNegocio = 'escrituraNegocio',
    EliminarNegocio = 'eliminarNegocio',
    AdminProducto = 'adminProducto',
    LecturaProducto = 'lecturaProducto',
    EscrituraProducto = 'escrituraProducto',
    EliminarProducto = 'eliminarProducto',
    AdminReview = 'adminReview',
    LecturaReview = 'lecturaReview',
    EscrituraReview = 'escrituraReview',
    EliminarReview = 'eliminarReview',
    AdminEvento = 'adminEvento',
    LecturaEvento = 'lecturaEvento',
    EscrituraEvento = 'escrituraEvento',
    EliminarEvento = 'eliminarEvento',
    AdminPago = 'adminPago',
    LecturaPago = 'lecturaPago',
    EscrituraPago = 'escrituraPago',
    EliminarPago = 'eliminarPago',
    AdminPedido = 'adminPedido',
    LecturaPedido = 'lecturaPedido',
    EscrituraPedido = 'escrituraPedido',
    EliminarPedido = 'eliminarPedido',
    AdminUsuario = 'adminUsuario',
    LecturaUsuario = 'lecturaUsuario',
    EscrituraUsuario = 'escrituraUsuario',
    EliminarUsuario = 'eliminarUsuario',
    AdminPublicacion = 'adminPublicacion',
    LecturaPublicacion = 'lecturaPublicacion',
    EscrituraPublicacion = 'escrituraPublicacion',
    EliminarPublicacion = 'eliminarPublicacion',
  }